Clean up password forms.
